Feat: 1. instantiate in Prefab is supported in Prefeb Manage tool
Refactor:
1. compress the skeleton relative tools
2. re-write descriptions for some tools
