+++ Editor/Scripts/API/Tool/GameObject.Skeleton.Analyze.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a44aa438fbb0a0d4fa56b9b7f4500239
+guid: 6fe58d58f0393b94c83ff698d536a850
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
