--- Editor/Scripts/API/Tool/GameObject.Skeleton.GetReferences.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8d127f7377a7f734cbe36694f33b16b7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
