--- Editor/Scripts/API/Tool/GameObject.Skeleton.StandardizeNaming.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9057a694bff8e834d9fecf3e35cc13a9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
